Refactor: Restructure project using app factory and blueprints
