+++ app/database.py
@@ -2,10 +2,7 @@ from flask_login import UserMixin
 from flask_sqlalchemy import SQLAlchemy 
 from werkzeug.security import generate_password_hash, check_password_hash # type: ignore
 from datetime import datetime
-
-# Initialize the database
-db = SQLAlchemy()
-
+from app.extensions import db
 
 # Friendship association table
 friendships = db.Table('friendships',
